Supporting using zenddevops/client as a vendor.
When installing zenddevops/client as a vendor into another project, it
adds the zs-client.php inside of vendor/bin.  When you try to run the
command...

php vendor/bin/zs-client.php installApp --zpk="/tmp/app-1.0.0.zpk"

...it could not find vendor/autoload.php.

This commit adds the cwd to the include path before it is changed inside
of zs-client.php.