Scripts directory within the ZPK was not being generated correctly for two reasons:
1) The files were contained within a double sub directory : /scripts/scripts
2) Sub-directories of the scripts folder were not being flattened, as is the case when packaging using Zend Studio

Packaging through Zend Studio allows directories to be included in scriptsdir.includes, the behaviour you see there is it copies all the files within that directory, along with all files in any sub-directories, into the scriptsdir defined in the XML - giving you a single directory in the package (named after scriptsdir) which contains all the files, flattened out with no sub-directories.

My update should match this behaviour, but I didnt check adding single files to the scriptsdir.includes as this looked to be handled fine already in the pack() method.